OpenSourceModels/FlashGeminiBot-main/FlashGeminiBot-main/requirements.txt
streamlit
langchain
sentence-transformers
uvicorn
ctransformers
python-box
langchain_community
google_generativeai
textwrap3
pathlib
ipython